Create Merging Notes and finalize continue-graphizing
This commit finalizes the design for continue-graph and creates some
test cases for the geb-gui.grahping:graphize function.

This PR also creates an idea of a merging note, which specifies that
the note should take the place of the node to the left of it when we
wish to continue drawing the graph.

This feature is somewhat complicated and thus tests have been made to
assure of its correctness

We've also added the gui tests to the testing framework so it can be
ran by CI. We should change this in the future potentnially