Simplify the graphing process massively by making multiple note types
Now notes come in both SQUASH-NOTE and NODE-NOTE, which represent
different kinds of forward facing notes.

A squash note is a potentially shared NOTE that will be the future
target.

A node note on the other hand is a NOTE that will be placed in the
graph itself rather than being squashed onto the note that comes
before it.

By moving out these two componenets the design is massively
sipmlified, and the logic fairly straight forward

With this work is the completion of the graphing function as well,
covering the morphism cases that were previous errors