Remove alias as an explicit type
Aliases now exist in the state that they are tagged to a specific
allocation. It will display the name, but it will not require special
care anymore.

This is not the final design, as a final design ought to be able to
recognize aliases where they occur rather than just tag a specific
instance of the idea. Opaque types should replace our aliases on
false, true etc.