Added Bitc polynomial backend and transformation to vampir
Adds category of boolean circuits (BITC). The objects in this category
are just natural numbers representing the type of bit vectors of a
certain length. There are nine methods for constructing morphisms;

- (compose x y), which composes morphisms x and y.
- (ident n), which is the identity on n.
- (fork n), which maps n onto 2*n by copying its inputs.
- (parallel x y), which (if x : a -> b and y : c -> d) will be a
  morphism from a + c -> b + d, running x and y on subvectors.
- (swap n m), which maps n + m onto m + n by swapping.
- ONE, which represents the map from 0 onto 1 producing a vector with only 1 in it.
- ZERO, which represents the map from 0 onto 1 producing a vector with only 0 in it.
- (drop n), which represents the unique morphism from n to 0.
- (branch x y), which (if x : a -> b and y : a -> b) maps 1+a to b by
  splitting on the first bit to decide which morphism to apply to the
  remaining bits.

There are other ways to formulate this category, but I think this
particular formulation is quite convenient.

I've implemented a `to-bitc` function in geb.trans which translates
geb objects and morphisms into bitc objects and
morphisms. Additionally, I've implemented a `to-vampir` function in
bitc.trans which translates a bit morphism into a vampire morphism.

I'm not sure what else is needed, but the core implementation is done
for now. In the future, this should be extended to a category whose
objects represent vectors over some finite set other than the
booleans. The reason I didn't do that hear is because coproducts are
only binary and there aren't finite sets beyond so0 and so1, so
bitvectors are quite natural and using anything else would require
post-hoc optimization, but future versions of geb may want more. Also,
I'd like to know what, if any, performance benefits this gives over
the univariate polynomial formulation. I didn't test that.