New VampIR Functions
Upgrade the Vamp-IR package to include basic functions of the library
such as checking whether something is zero, checking whether a number
is negative etc. Alongside that, arithmetic operations with appropriate range
checks are introduced in order to be used in the upgraded pipeline
which will allow for the compilation of natural number functions and
predicates of fixed bitwidth.