API for Checking Well-Definedness of Morphisms
Introduces a well-defp-cat function to Geb and SeqN which takes in a
morphism in the appropriate category and either gives true if the
morphism is well-defined or produces an error message pointing to the
point at which the moprhism fails to be well-defined, e.g. appropriate
domains and codomains do not coincide for composition of morphisms.