New SeqN Category and API
Introduces a new category SeqN with appropriate API alongside with
the compilation of it to appropriate Vamp-IR code. SeqN can be
thought of as a category of finite sequences of natural numbers
with morphisms from (x1,...,xn) to (y1,...,ym) being Vamp-IR
functions taking in n entries of sizes x1 to xn and spitting out
m outputs of sizes y1 to ym.