Introdution of Maybe wrapper
Changes the compilation of Lambda terms with errors to one using the
actual Maybe wrapper, instead of a recursive one. In return, we ask
for the error term to not have any function application or lambda
terms appearing outside of the top of the term.

The wrapper assumer 24-bit integer inputs as per Juvix testing specification.