Fix to-circuit Variable Printing
Previously a Lambda function taking natural number arguments `A1` to
`An` would be compiled to a circuit with arguments `x1` to `x(n+1)`
where argument `xi` in the circuit for i<(n+1) stands for argument
`A(n-i)` and `x(n+1)` stands as an unused variable.

Current channge omits the final variable and renames the rest, so
that the compiled circuit will have appropriate argument order and
type corresponding to the relevant STLC term.

Note that this changes how to input coproduct type circuit variables.

A